requirements.txt
schedule
mysql-connector-python-rf
requests